Оно работает!!!

+++ src/main/java/ru/sgu/hell/dh/model/CryptGenerator.java
@@ -1,7 +1,5 @@
 package ru.sgu.hell.dh.model;
 
-import java.math.BigInteger;
-
 public interface CryptGenerator {
-	BigInteger generate();
+	Long generate();
 }
